move segment related stuff to separate file
